Update Client.h
Changing the sign of the W_n_k and W in the FFT algorithm. This change does not affect the magnitude of the data, but it flips the phase. This change is not reflected in the executable file, since the phase is not plotted. The user will have to recompile the code if he/she wants to have this change reflected in the executable file.